Mention inter subspec dependencies
I faced this issue when trying to push the publish the pod to CocoaPods
-- the `pod trunk push` command started failing at the linting step.

From reading the docs (http://www.rdoc.info/github/CocoaPods/Core/Pod/Specification/DSL:subspec) I think this would be the missing link, but I'm not sure if this introduces any other issues or not. FWIW,

    pod spec lint TDTChocolate.podspec

works after this change. We'll know the behaviour of `pod trunk push`
only after we get this merged.